feat: add bot integration banner
