+++ locales/en-US/common.json
@@ -28,6 +28,8 @@
   "back": "Back",
   "batchDelete": "Batch Delete",
   "blog": "Product Blog",
+  "botIntegrationBanner.dismiss": "Dismiss",
+  "botIntegrationBanner.title": "Add Channels to LobeAI",
   "branching": "Create Subtopic",
   "branchingDisable": "The \"Sub-topic\" feature is unavailable in the current mode. To use this feature, please switch to Postgres/Pglite DB mode or use LobeHub Cloud.",
   "branchingRequiresSavedTopic": "Current topic is not saved, please save it first to use subtopic feature",
