+++ locales/en-US/plugin.json
@@ -617,8 +617,6 @@
   "skillDetail.tabs.tools": "Capabilities",
   "skillDetail.tools": "Tools",
   "skillDetail.trustWarning": "Only use connectors from developers you trust. LobeHub does not control which tools developers make available and cannot verify that they will work as intended or that they won't change.",
-  "skillInstallBanner.dismiss": "Dismiss",
-  "skillInstallBanner.title": "Add skills to Lobe AI",
   "store.actions.cancel": "Cancel",
   "store.actions.configure": "Configure",
   "store.actions.confirmUninstall": "Uninstalling will clear Skill config. Continue?",
