+++ src/locales/default/common.ts
@@ -31,6 +31,8 @@ export default {
   'back': 'Back',
   'batchDelete': 'Batch Delete',
   'blog': 'Product Blog',
+  'botIntegrationBanner.dismiss': 'Dismiss',
+  'botIntegrationBanner.title': 'Add Channels to LobeAI',
   'branching': 'Create Subtopic',
   'branchingDisable':
     'The "Sub-topic" feature is unavailable in the current mode. To use this feature, please switch to Postgres/Pglite DB mode or use LobeHub Cloud.',
